pythonProject/requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
cryptography>=41.0.0
psutil>=5.9.0